requirements.txt
affine==2.3.0
Fiona==1.8.22
geogif==0.3
geopandas==0.12.0
imageio==2.35.1
matplotlib==3.4.3
netCDF4==1.7.2
numpy==1.22.4
pandas==1.3.4
PyYAML==6.0
rasterio==1.2.10
rioxarray==0.12.2
ruamel.base==1.0.0
scikit_learn==1.0.1
Shapely==2.0.7
torchmetrics==1.5.0
tqdm==4.62.3
xarray==0.20.1
